+++ CopilotKit/examples/next-openai/src/app/api/copilotkit/langchain/route.ts
@@ -30,8 +30,8 @@ export async function POST(req: Request): Promise<Response> {
   return copilotKit.response(
     req,
     new LangChainAdapter(async (forwardedProps) => {
-      const model = new ChatOpenAI({ modelName: "gpt-4" });
-      return model.stream(forwardedProps.messages, { functions: forwardedProps.functions });
+      const model = new ChatOpenAI({ modelName: "gpt-4-1106-preview" });
+      return model.stream(forwardedProps.messages, { tools: forwardedProps.tools });
     }),
   );
 }