+++ CopilotKit/packages/react-core/src/lib/copilot-task.ts
@@ -65,7 +65,7 @@ export class CopilotTask<T = any> {
     const response = await fetchAndDecodeChatCompletion({
       copilotConfig: context.copilotApiConfig,
       messages: messages,
-      functions: context.getChatCompletionFunctionDescriptions(entryPoints),
+      tools: context.getChatCompletionFunctionDescriptions(entryPoints),
       headers: context.copilotApiConfig.headers,
       body: context.copilotApiConfig.body,
     });
@@ -75,7 +75,7 @@ export class CopilotTask<T = any> {
     }
 
     const reader = response.events.getReader();
-    let functionCall: FunctionCall | undefined;
+    let functionCalls: FunctionCall[] = [];
 
     while (true) {
       const { done, value } = await reader.read();
@@ -85,20 +85,22 @@ export class CopilotTask<T = any> {
       }
 
       if (value.type === "function") {
-        functionCall = {
+        functionCalls.push({
           name: value.name,
           arguments: JSON.stringify(value.arguments),
-        };
+        });
         break;
       }
     }
 
-    if (!functionCall) {
+    if (!functionCalls.length) {
       throw new Error("No function call occurred");
     }
 
     const functionCallHandler = context.getFunctionCallHandler(entryPoints);
-    await functionCallHandler(messages, functionCall);
+    for (const functionCall of functionCalls) {
+      await functionCallHandler(messages, functionCall);
+    }
   }
 }