+++ CopilotKit/packages/shared/src/types/openai-assistant.ts
@@ -37,7 +37,7 @@ export interface Message {
   function_call?: string | FunctionCall;
 }
 
-export interface Function {
+export interface FunctionDefinition {
   /**
    * The name of the function to be called. Must be a-z, A-Z, 0-9, or contain
    * underscores and dashes, with a maximum length of 64.
@@ -58,6 +58,11 @@ export interface Function {
    * how to call the function.
    */
   description?: string;
+}
+
+export interface ToolDefinition {
+  type: "function";
+  function: FunctionDefinition;
 }
 
 export type FunctionCallHandler = (