+++ CopilotKit/packages/react-textarea/CHANGELOG.md
@@ -1,5 +1,17 @@
 # ui
 
+## 0.30.0-tools.0
+
+### Minor Changes
+
+- Test the tools API
+
+### Patch Changes
+
+- Updated dependencies
+  - @copilotkit/react-core@0.20.0-tools.0
+  - @copilotkit/shared@0.4.0-tools.0
+
 ## 0.29.0
 
 ### Minor Changes