+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @copilotkit/shared
 
+## 0.4.0-tools.0
+
+### Minor Changes
+
+- Test the tools API
+
 ## 0.3.0
 
 ### Minor Changes